Enhance Campaigns API with Filtering Options by Type and Category
Improve the getAll method in the CampaignsController to support filtering campaigns by type (CURRENT/UPCOMING) and category_id. If no filters are specified, all campaigns will be returned. The getCampaigns method in the CampaignsService has been updated to handle the new filters and retrieve the filtered campaigns from the database.

Related to #4